Relocated firebase files to lib/configs
