+++ app/api/auth/register-admin/route.js
@@ -1,5 +1,5 @@
 import { registerUser } from '@/services/authService';
-import { db } from '@/lib/firebase';
+import { db } from '@/lib/configs/firebase';
 import { successResponse, errorResponse } from '@/utils/response';
 
 export async function POST(req) {
