+++ components/modals/EditSpotModal.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { auth } from "@/lib/firebaseClient";
+import { auth } from "@/lib/configs/firebaseClient";
 import { Button } from "@/components/ui/button";
 
 type SensorInfo = {
