+++ lib/configs/firebase.js
