+++ lib/configs/firebaseClient.js
