+++ services/logout.ts
@@ -1,5 +1,5 @@
 import { signOut } from "firebase/auth";
-import { auth } from "@/lib/firebaseClient";
+import { auth } from "@/lib/configs/firebaseClient";
 
 export const logoutUser = async () => {
   await signOut(auth);
