+++ services/spotService.js
@@ -5,7 +5,7 @@
 //          /api/spots/[slotId]
 // ============================================================
 
-import { db } from '@/lib/firebase';
+import { db } from '@/lib/configs/firebase';
 import { v4 as uuidv4 } from 'uuid';
 
 const VALID_STATUSES = ['available', 'occupied', 'reserved', 'disabled'];
